requirements.txt
PyJWT
asgiref==3.2.3
Django==3.0.3
pytz==2019.3
sqlparse==0.3.0
django-cors-headers==3.2.1